refactoring new packaging strategy
