--- src/main/java/constants/Exceptions.java
@@ -1,6 +1,0 @@
-package constants;
-
-// TODO: WORK IN PROGESS. Make it so we fail cleanly :)
-
-public class Exceptions {
-}
